+++ flink-python/src/main/java/org/apache/flink/streaming/api/runners/python/beam/PythonFunctionRunner.java
@@ -16,10 +16,11 @@
  * limitations under the License.
  */
 
-package org.apache.flink.python;
+package org.apache.flink.streaming.api.runners.python.beam;
 
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.api.java.tuple.Tuple2;
+import org.apache.flink.python.PythonConfig;
 
 /** The base interface of runner which is responsible for the execution of Python functions. */
 @Internal
